+++ android/app/src/main/java/com/logseq/og/FsWatcher.java
@@ -1,4 +1,4 @@
-package com.logseq.app;
+package com.logseq.og;
 
 import android.system.ErrnoException;
 import android.system.Os;
