+++ android/app/src/main/java/com/logseq/og/MainActivity.java
@@ -1,4 +1,4 @@
-package com.logseq.app;
+package com.logseq.og;
 
 import android.content.Intent;
 import android.os.Bundle;
