+++ android/app/src/main/res/values/strings.xml
@@ -2,6 +2,6 @@
 <resources>
     <string name="app_name">Logseq OG</string>
     <string name="title_activity_main">Logseq OG</string>
-    <string name="package_name">com.logseq.app</string>
-    <string name="custom_url_scheme">com.logseq.app</string>
+    <string name="package_name">com.logseq.og</string>
+    <string name="custom_url_scheme">com.logseq.og</string>
 </resources>
