+++ ios/App/App/App.entitlements
@@ -6,7 +6,7 @@
 	<string>development</string>
 	<key>com.apple.developer.icloud-container-identifiers</key>
 	<array>
-		<string>iCloud.com.logseq.logseq</string>
+		<string>iCloud.com.logseq.og</string>
 	</array>
 	<key>com.apple.developer.icloud-services</key>
 	<array>
@@ -14,11 +14,11 @@
 	</array>
 	<key>com.apple.developer.ubiquity-container-identifiers</key>
 	<array>
-		<string>iCloud.com.logseq.logseq</string>
+		<string>iCloud.com.logseq.og</string>
 	</array>
 	<key>com.apple.security.application-groups</key>
 	<array>
-		<string>group.com.logseq.logseq</string>
+		<string>group.com.logseq.og</string>
 	</array>
 </dict>
 </plist>
