+++ ios/App/App/AppDebug.entitlements
@@ -8,7 +8,7 @@
 	<string>Development</string>
 	<key>com.apple.developer.icloud-container-identifiers</key>
 	<array>
-		<string>iCloud.com.logseq.logseq</string>
+		<string>iCloud.com.logseq.og</string>
 	</array>
 	<key>com.apple.developer.icloud-services</key>
 	<array>
@@ -16,11 +16,11 @@
 	</array>
 	<key>com.apple.developer.ubiquity-container-identifiers</key>
 	<array>
-		<string>iCloud.com.logseq.logseq</string>
+		<string>iCloud.com.logseq.og</string>
 	</array>
 	<key>com.apple.security.application-groups</key>
 	<array>
-		<string>group.com.logseq.logseq</string>
+		<string>group.com.logseq.og</string>
 	</array>
 </dict>
 </plist>
