+++ ios/App/ShareViewController/ShareViewController.entitlements
@@ -4,7 +4,7 @@
 <dict>
 	<key>com.apple.security.application-groups</key>
 	<array>
-		<string>group.com.logseq.logseq</string>
+		<string>group.com.logseq.og</string>
 	</array>
 </dict>
 </plist>
